a2a-plaground/client-agent/pyproject.toml
[project]
name = "client-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk[a2a]>=1.6.1",
]
